Fixes a glitch of opening editor at first click
If you draw more than three polygons, the editor was opening in the first click of the third polygon.

It was happening because Annotorious is sending two clicks in the second polygon, three clicks in the third, and so on. The isClosed function checks if it's at least the third point and if it's 5 pixels close to the _anchor point, so it was causing the window to open wrongly.

My contribution is a workaround to check if the click is the same as the last one, so it doesn't call the isClosed function if so.

It doesn't correct the multiple-click event issue (I think it's a core Annotorious glitch), but it corrects the expected behavior.